Fix Page Header Being Too Bold
Fix the page header being too bold in the user section of the page.
The header was looking right on the home page of the admin area and on
the preview page, but not on the regular page.
Fixed through inclusion of bootstrap and its dependencies on the normal
page.

Signed-off-by: ahahn94 <ahahn94@outlook.com>